requirements.txt
feedparser>=6.0.0
openai>=1.12.0
requests>=2.31.0
streamlit>=1.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=5.1.0
lxml_html_clean>=0.1.0
